+++ doc/changelog.md
@@ -10,7 +10,7 @@ Syntax: `- short text describing the change _(Your Name)_`
 
 - _()_
 - _()_
-- _()_
+- Exit planting mode with escape _(Daniel Steinkogler)_
 - Renaming of layer list components _(Daniel Steinkogler)_
 - Fix ordering of seeds by expiration date _(Moritz)_
 - Re-enable seed search. It will now take the entire plant name into account. _(Moritz)_